+++ backend/prisma/migrations/20260330202644_nullable_type_fields/migration 2.sql
@@ -1,0 +1,5 @@
+-- AlterTable
+ALTER TABLE "Group" ALTER COLUMN "type" DROP NOT NULL;
+
+-- AlterTable
+ALTER TABLE "Location" ALTER COLUMN "type" DROP NOT NULL;